Drop unneeded dependency on the `windows` crate & further cleanups (#15)
* Extend `windows` dependency range to `0.58-0.61`

* cargo: Set `rust-version` to keep in sync with the CI MSRV

Set the `rust-version` in `Cargo.toml` for a multitude of reasons
(crates.io shows it to users, and `cargo` can use it in its MSRV-aware
resolver), including the `+nightly -Zminimal-versions generate-lockfile`
step to generate a compatible `version = 3` `Cargo.lock` for MSRV 1.74
to consume, rather than `version = 4` which it doesn't understand yet.

* Fix clippy lints

* Minor code and API cleanups

* Drop unneeded dependency on the `windows` crate

This saves yet one more crate to update, because it was only used to
query the size of the `LUID` type which is always known to be 8 bytes.

If anything the more light-weight `windows_sys` crate should have been
used instead.

* CI: Run on all OS'es

* CI: Diff generated result from `api_gen`